add parameterization to algorithm and scripts
Added parameterization to fineman's algorithm.
Currently it defaults to float, but it is possible
to change to Decimal and int.

Graph generator scripts have also been changed
accordingly to allow generation of different types.

Currently we only do tests on int and floats and
make a sanity check on random graphs with no negative
cycles on all three weight types.